Add generating 3D volumes and masks from segmentations
Also generate a json with dicom metadata of the first dcm file
and a list of slices, with same order than the volume and SOPInstanceUID

When create_volumes and download_dicoms are ON, the grayscale
volume will be created in the downloaded dicoms folder.
The masks will be generated in the segmentations folder.

Improve README